+++ src/app/suppliers/suppliers.component.ts
@@ -4,6 +4,7 @@ import {MatTableDataSource} from "@angular/material/table";
 import {Supplier} from "../api/models/supplier";
 import {Router} from "@angular/router";
 import {SupplierService} from "./suppliers.service";
+import {FuseConfirmationService} from "../../@fuse/services/confirmation";
 
 @Component({
   selector: 'app-suppliers',
@@ -18,7 +19,8 @@ export class SuppliersComponent implements OnInit{
     displayedColumns = ['id', 'name', 'contactPerson', 'editDelete'];
 
     constructor(private supplierService: SupplierService,
-                private router: Router,) {
+                private router: Router,
+                private fuseConfirmationService: FuseConfirmationService) {
 
     }
 
@@ -36,8 +38,24 @@ export class SuppliersComponent implements OnInit{
     }
 
     deleteSupplier(supplierId: string){
-        this.supplierService.deleteSupplier(supplierId).pipe(takeUntil(this.endsubs$)).subscribe(() => {
-            this._getSuppliers();
+        const confirmation = this.fuseConfirmationService.open({
+            title: 'Delete supplier',
+            message: 'Are you sure you want to remove this supplier? This action cannot be undone!',
+            actions: {
+                confirm: {
+                    label: 'Delete'
+                }
+            }
+        });
+
+        confirmation.afterClosed().subscribe((result) => {
+
+            // If the confirm button pressed...
+            if (result === 'confirmed') {
+                this.supplierService.deleteSupplier(supplierId).pipe(takeUntil(this.endsubs$)).subscribe(() => {
+                    this._getSuppliers();
+                });
+            }
         });
     }
